tSNE/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.2.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.9.0
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
